Create TimelineWriterTester to abstract SVG and HTML tests
